+++ example/example/spiders/books_04_overrides_02.py
@@ -41,7 +41,7 @@ class BTSBookPage(BookPage):
 class BPBookListPage(BookListPage):
     """Logic to extract listings from pages like https://bookpage.com/reviews"""
     def book_urls(self):
-        return self.css('.article-info a::attr(href)').getall()
+        return self.css('article.post h4 a::attr(href)').getall()
 
 
 class BPBookPage(BookPage):
@@ -49,7 +49,7 @@ class BPBookPage(BookPage):
     def to_item(self):
         return {
             'url': self.url,
-            'name': self.css(".book-data h4::text").get().strip(),
+            'name': self.css("body div > h1::text").get().strip(),
         }
 
 
@@ -58,16 +58,12 @@ class BooksSpider(scrapy.Spider):
     start_urls = ['http://books.toscrape.com/', 'https://bookpage.com/reviews']
     # Configuring different page objects pages for different domains
     custom_settings = {
-        "SCRAPY_POET_OVERRIDES": {
-            "toscrape.com": {
-                BookListPage: BTSBookListPage,
-                BookPage: BTSBookPage
-            },
-            "bookpage.com": {
-                BookListPage: BPBookListPage,
-                BookPage: BPBookPage
-            },
-        }
+        "SCRAPY_POET_OVERRIDES": [
+            ("toscrape.com", BTSBookListPage, BookListPage),
+            ("toscrape.com", BTSBookPage, BookPage),
+            ("bookpage.com", BPBookListPage, BookListPage),
+            ("bookpage.com", BPBookPage, BookPage)
+        ]
     }
 
     def parse(self, response, page: BookListPage):