+++ scrapy_poet/utils.py
@@ -1,21 +1,6 @@
 import os
 
 from scrapy.utils.project import project_data_dir, inside_project
-from tldextract import tldextract
-
-
-def get_domain(url):
-    """
-    Return the domain without any subdomain
-
-    >>> get_domain("http://blog.example.com")
-    'example.com'
-    >>> get_domain("http://www.example.com")
-    'example.com'
-    >>> get_domain("http://deeper.blog.example.co.uk")
-    'example.co.uk'
-    """
-    return ".".join(tldextract.extract(url)[-2:])
 
 
 def get_scrapy_data_path(createdir=True):